changed delete user and password change endpoints

+++ user/views.py
@@ -1,33 +1,29 @@
-from django.contrib.auth import (
-    login as django_login,
-    logout as django_logout
-)
+from allauth.account.utils import complete_signup
+from allauth.account import app_settings as allauth_settings
 
+from contact.models import Contact
 from django.conf import settings
-from django.core.exceptions import ObjectDoesNotExist
-from django.contrib.auth.models import User
-
+from django.contrib.auth import (
+    hashers as hashers
+)
 from django.utils.decorators import method_decorator
 from django.utils.translation import ugettext_lazy as _
 from django.views.decorators.debug import sensitive_post_parameters
 
 from rest_framework import status
-from rest_framework.views import APIView
 from rest_framework.response import Response
-from rest_framework.generics import GenericAPIView, CreateAPIView
-from rest_framework.permissions import AllowAny, IsAuthenticated
+from rest_framework.generics import CreateAPIView
+from rest_framework.permissions import IsAuthenticated
 
-from allauth.account.utils import complete_signup
-from allauth.account import app_settings as allauth_settings
+from pymongo import MongoClient
 
 from .app_settings import (
-    LoginSerializer, TokenSerializer, JWTSerializer,
-    RegisterSerializer, create_token, register_permission_classes,
+    TokenSerializer, JWTSerializer, RegisterSerializer,
+    create_token, register_permission_classes
 )
 
 from .models import TokenModel
 from .utils import jwt_encode
-
 
 
 sensitive_post_parameters_m = method_decorator(
@@ -35,99 +31,6 @@ sensitive_post_parameters_m = method_decorator(
         'password', 'old_password', 'new_password1', 'new_password2', 'password1', 'password2'
     )
 )
-
-class LoginView(GenericAPIView):
-    """
-    Check the credentials and return the REST Token
-    if the credentials are valid and authenticated.
-    Calls Django Auth login method to register User ID
-    in Django session framework
-
-    Accept the following POST parameters: username, password
-    Return the REST Framework Token Object's key.
-    """
-    permission_classes = (AllowAny,)
-    serializer_class = LoginSerializer
-    token_model = TokenModel
-
-    @sensitive_post_parameters_m
-    def dispatch(self, *args, **kwargs):
-        return super(LoginView, self).dispatch(*args, **kwargs)
-
-    def process_login(self):
-        django_login(self.request, self.user)
-
-    def get_response_serializer(self):
-        if getattr(settings, 'REST_USE_JWT', False):
-            response_serializer = JWTSerializer
-        else:
-            response_serializer = TokenSerializer
-        return response_serializer
-
-    def login(self):
-        self.user = self.serializer.validated_data['user']
-
-        if getattr(settings, 'REST_USE_JWT', False):
-            self.token = jwt_encode(self.user)
-        else:
-            self.token = create_token(self.token_model, self.user,
-                                      self.serializer)
-
-        if getattr(settings, 'REST_SESSION_LOGIN', True):
-            self.process_login()
-
-    def get_response(self):
-        serializer_class = self.get_response_serializer()
-
-        if getattr(settings, 'REST_USE_JWT', False):
-            data = {
-                'user': self.user,
-                'token': self.token
-            }
-            serializer = serializer_class(instance=data,
-                                          context={'request': self.request})
-        else:
-            serializer = serializer_class(instance=self.token,
-                                          context={'request': self.request})
-
-        return Response(serializer.data, status=status.HTTP_200_OK)
-
-    def post(self, request, *args, **kwargs):
-        self.request = request
-        self.serializer = self.get_serializer(data=self.request.data,
-                                              context={'request': request})
-        self.serializer.is_valid(raise_exception=True)
-
-        self.login()
-        return self.get_response()
-
-class LogoutView(APIView):
-
-    permission_classes = (AllowAny,)
-
-    def get(self, request, *args, **kwargs):
-        if getattr(settings, 'ACCOUNT_LOGOUT_ON_GET', False):
-            response = self.logout(request)
-        else:
-            response = self.http_method_not_allowed(request, *args, **kwargs)
-
-        return self.finalize_response(request, response, *args, **kwargs)
-
-    def post(self, request, *args, **kwargs):
-        return self.logout(request)
-
-    def logout(self, request):
-        try:
-            request.user.auth_token.delete()
-        except (AttributeError, ObjectDoesNotExist):
-            pass
-
-        django_logout(request)
-
-        return Response({"detail": _("Successfully logged out.")},
-                        status=status.HTTP_200_OK)
-
-
 class RegisterView(CreateAPIView):
     serializer_class = RegisterSerializer
     permission_classes = register_permission_classes()
@@ -177,30 +80,31 @@ class RegisterView(CreateAPIView):
 class DeleteUserView(CreateAPIView):
     permission_classes = (IsAuthenticated,)
 
+    def get_owner(self, request):
+        client = MongoClient('localhost', 27017)
+        users = client.TaggedDB.auth_user
+        user = users.find_one({"id": request.user.id})
+        return str(user['_id'])
+
+    def get_password(self, request):
+        client = MongoClient('localhost', 27017)
+        users = client.TaggedDB.auth_user
+        user = users.find_one({"id": request.user.id})
+        return str(user['password'])
+
     def get(self, request):
+        owner = self.get_owner(request)
+        contacts = Contact.objects.all().filter(owner=owner)
+        contacts.delete()
         request.user.delete()
         return Response("User deleted.", status=204)
 
     def post(self, request):
-        request.user.delete()
-        return Response("Something went wrong. User not deleted,", status=204)
-
-
-class ChangePasswordView(CreateAPIView):
-    permission_classes = (IsAuthenticated,)
-    serializer_class = LoginSerializer
-
-    def post(self, request):
-        serializer_class = PasswordChangeSerializer
-    permission_classes = (IsAuthenticated,)
-
-    @sensitive_post_parameters_m
-    def dispatch(self, *args, **kwargs):
-        return super(ChangePasswordView, self).dispatch(*args, **kwargs)
-
-    def post(self, request, *args, **kwargs):
-        serializer = self.get_serializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        serializer.save()
-        return Response({"detail": _("New password has been saved.")})
-
+        if hashers.check_password(request.data.get('password'), self.get_password(request)):
+            owner = self.get_owner(request)
+            contacts = Contact.objects.all().filter(owner=owner)
+            contacts.delete()
+            request.user.delete()
+            return Response("", status=204)
+        else:
+            return Response("", status=500)
